Moved more functions out of client
