+++ webAO/client/isLowMemory.ts
@@ -1,0 +1,10 @@
+import { setOldLoading } from '../client'
+export const isLowMemory = () => {
+    if (
+        /webOS|iPod|BlackBerry|BB|PlayBook|IEMobile|Windows Phone|Kindle|Silk|PlayStation|Nintendo|Opera Mini/i.test(
+            navigator.userAgent
+        )
+    ) {
+        setOldLoading(true);
+    }
+}
